refactor: sink model cost resolution into the pricing module
Move CostSource and resolve_model_cost out of the usage feature into
pricing, and collapse the duplicate merged-row cost resolver so usage
and display share one implementation.
